add entity: delivery: Delivery, order: OrderDetail, OrderStatus, Product: Category, ProductImage, VarientAttribute, storage: Inventory, Warehouse, StockMovement; refactor packages: cart, product, storage, user: User, puml -> changes String IDs to int IDs

+++ src/main/java/viettech/entity/user/User.java
@@ -11,7 +11,7 @@ public abstract class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_id")
-    protected Long userId;
+    protected int userId;
 
     @Column(name = "first_name", nullable = false)
     protected String firstName;
